Extract toJson/fromJson into separate class: TransactionEncoder
