+++ packages/iov-core/src/index.ts
@@ -28,3 +28,4 @@ export { JsRpcSigningServer } from "./jsrpcsigningserver";
 export { JsonRpcSigningServer } from "./jsonrpcsigningserver";
 export { MultiChainSigner } from "./multichainsigner";
 export { GetIdentitiesAuthorization, SignAndPostAuthorization, SigningServerCore } from "./signingservercore";
+export { TransactionEncoder } from "./transactionencoder";
