+++ packages/iov-core/types/index.d.ts
@@ -5,3 +5,4 @@ export { JsRpcSigningServer } from "./jsrpcsigningserver";
 export { JsonRpcSigningServer } from "./jsonrpcsigningserver";
 export { MultiChainSigner } from "./multichainsigner";
 export { GetIdentitiesAuthorization, SignAndPostAuthorization, SigningServerCore } from "./signingservercore";
+export { TransactionEncoder } from "./transactionencoder";
